Add files via upload
Decision Tree: base don the root and the hirerachy it will create a tree, and we have to build this tree only on independent variable, Gini, entropy max leaf minimum sample are some of the hyper parameter or prunning techniques we will use in Decision Tree. Purnning means trimming the tree it will reduce the overfitting problem.